Add extendSubtask function
Users can click the '+ 1 min' button to add another minute to the
subtask they are currently on.

This function works by adding 1 to the value in the subtask's
input field
and then immediately restarting the timer.

I have also added a 'class' function to the input fields, which is used
to tag each input as incomplete or complete
so that when the timer restarts it only adds the tasks to go to the
new countdown.